Signedpoff-by: zhdy2009011286 <rvn76@qq.com> se of the "test_gethostname.c" is to test the correctness of the input hostname.If correct, the hostname will be returned by the function.
The purpose of the "test_gethostbyname.c" is to get the information of the host through the execution of the "gethostbyname" function with the name of the host as a string type parameter. If the function can be executed without any error, it will show the address family and the address of the host.